feat : restrict api to authenticated user

+++ src/api/folder.ts
@@ -6,14 +6,25 @@ import { z } from 'zod';
 import { revalidatePath } from 'next/cache';
 import { redirect } from 'next/navigation';
 import { unstable_noStore as noStore } from 'next/cache';
-import { USERS } from '../../scripts/placeholder';
+import { auth } from "../auth"
 
 export async function fetchRootFolders() {
   noStore();
+
+  const session = await auth()
+  if (!session) {
+    redirect("/auth/login")
+  }
+
+  const userId = session.user?.id;
+  if (!userId) {
+    throw new Error('Authentication Error: User not found.');
+  }
+
   try {
     const folders = await sql<Folder>`
       SELECT * FROM folders
-      WHERE folder_id IS NULL
+      WHERE folder_id IS NULL AND user_id = ${userId}
     `;
     return folders.rows;
   } catch (error) {
@@ -23,10 +34,21 @@ export async function fetchRootFolders() {
 }
 export async function fetchFolderById(id: string) {
   noStore();
+
+  const session = await auth()
+  if (!session) {
+    redirect("/auth/login")
+  }
+
+  const userId = session.user?.id;
+  if (!userId) {
+    throw new Error('Authentication Error: User not found.');
+  }
+
   try {
     const folders = await sql<Folder>`
       SELECT * FROM folders
-      WHERE id = ${id}
+      WHERE id = ${id} AND user_id = ${userId}
     `;
     return folders.rows[0];
   } catch (error) {
@@ -36,10 +58,21 @@ export async function fetchFolderById(id: string) {
 }
 export async function fetchFoldersByFolderId(id: string) {
   noStore();
+
+  const session = await auth()
+  if (!session) {
+    redirect("/auth/login")
+  }
+
+  const userId = session.user?.id;
+  if (!userId) {
+    throw new Error('Authentication Error: User not found.');
+  }
+
   try {
     const folders = await sql<Folder>`
       SELECT * FROM folders
-      WHERE folder_id = ${id}
+      WHERE folder_id = ${id} AND user_id = ${userId}
     `;
     return folders.rows;
   } catch (error) {
@@ -78,6 +111,16 @@ export type State = {
 
 export async function createFolder(folderId: string | null, prevState: State, formData: FormData) {
 
+  const session = await auth()
+  if (!session) {
+    redirect("/auth/login")
+  }
+
+  const userId = session.user?.id;
+  if (!userId) {
+    return { message: 'Authentication Error: User not found.' };
+  }
+
   // Validate form fields using Zod
   const validatedFields = CreateFolder.safeParse({
     title: formData.get('title'),
@@ -94,7 +137,6 @@ export async function createFolder(folderId: string | null, prevState: State, fo
 
   // Prepare data for insertion into the database
   const { title, description } = validatedFields.data;
-  const userId = USERS[0].id;
   // const date = new Date().toISOString().split('T')[0];
 
   let insertedId: string;
@@ -123,6 +165,16 @@ const UpdateFolder = FolderFormSchema.omit({ id: true/* , date: true */ });
 
 export async function updateFolder(folder: Folder, prevState: State, formData: FormData) {
 
+  const session = await auth()
+  if (!session) {
+    redirect("/auth/login")
+  }
+
+  const userId = session.user?.id;
+  if (!userId) {
+    return { message: 'Authentication Error: User not found.' };
+  }
+
   // Validate form fields using Zod
   const validatedFields = UpdateFolder.safeParse({
     title: formData.get('title'),
@@ -139,14 +191,13 @@ export async function updateFolder(folder: Folder, prevState: State, formData: F
 
   // Prepare data for insertion into the database
   const { title, description } = validatedFields.data;
-  const userId = USERS[0].id;
   // const date = new Date().toISOString().split('T')[0];
 
   try {
     await sql`
       UPDATE folders
       SET title = ${title}, description = ${description}
-      WHERE id = ${folder.id}
+      WHERE id = ${folder.id} AND user_id = ${userId}
     `;
   } catch (error) {
     return {
@@ -162,8 +213,18 @@ export async function updateFolder(folder: Folder, prevState: State, formData: F
 
 export async function deleteFolder(folder: Folder) {
 
+  const session = await auth()
+  if (!session) {
+    redirect("/auth/login")
+  }
+
+  const userId = session.user?.id;
+  if (!userId) {
+    return { message: 'Authentication Error: User not found.' };
+  }
+
   try {
-    await deleteFolderRecursively(folder);
+    await deleteFolderRecursively(folder, userId);
   } catch (error) {
     return { message: 'Database Error: Failed to Delete Folder.' };
   }
@@ -173,17 +234,17 @@ export async function deleteFolder(folder: Folder) {
 
 }
 
-async function deleteFolderRecursively(folder: Folder) {
+async function deleteFolderRecursively(folder: Folder, userId: string) {
 
   // delete all folder notes
-  await sql`DELETE FROM notes WHERE folder_id = ${folder.id}`;
+  await sql`DELETE FROM notes WHERE folder_id = ${folder.id} AND user_id = ${userId}`;
 
   const folders = await fetchFoldersByFolderId(folder.id);
 
   for (const folder of folders) {
-    await deleteFolderRecursively(folder);
+    await deleteFolderRecursively(folder, userId);
   }
 
-  await sql`DELETE FROM folders WHERE id = ${folder.id}`;
+  await sql`DELETE FROM folders WHERE id = ${folder.id} AND user_id = ${userId}`;
 
 }
